\name{biotransformationsTable}
\alias{biotransformationsTable}
\title{biotransformationsTable}
\description{
This table contains the biotransformations to be looked up for.
}
\value{
A table having the fields:
\itemize{
\item{
NAME: The name of the biotransformation
}
\item{
MASSDIFF: The mass difference of the fragment caused by the biotransformation
}
}
}
\seealso{
  \code{\link{Biotransformations}}}

\author{Francesc Fernandez,
 \email{francesc.fernandez.albert@upc.edu}}
\keyword{file}
