\name{Database}
\alias{Database}
\title{Human Metabolome Database}
\description{
The Human Metabolome Database is saved in this dataframe.
}
\value{
A table having the fields:
\itemize{
\item{
ENTRY: HMDB entry
}
\item{
NAME: The compound name
}
\item{
FORMULA: The chemical formula of the compound
}
\item{
MASS: Mass of the fragment
}
\item{
Biofluid: Where the compound can be found
}
}
}
\seealso{
  \code{\link{identifyMetabolites}}  \code{\link{metaboliteTable}}}

\author{Francesc Fernandez,
 \email{francesc.fernandez.albert@upc.edu}}
\keyword{file}
