\name{MAIT.Parameters-class}
\Rdversion{1.1}
\docType{class}
\alias{MAIT.Parameters-class}

\title{Class \code{"MAIT.Parameters"}}
\description{
This class contains all the parameters used in the MAIT run.
}
\section{Slots}{
  \describe{
    \item{\code{sampleProcessing}:}{List containing the parameters of the function \link{sampleProcessing}}
    \item{\code{peakAnnotation}:}{List containing the parameters of the function \link{peakAnnotation}}
    \item{\code{peakAggregation}:}{List containing the parameters of the function \link{peakAggregation}}
    \item{\code{sigFeatures}:}{List containing the parameters of the function \link{spectralSigFeatures}}
    \item{\code{biotransformations}:}{List containing the parameters of the function \link{Biotransformations}}
    \item{\code{identifyMetabolites}:}{List containing the parameters of the function \link{identifyMetabolites}}
    \item{\code{classification}:}{List containing the parameters of the function \link{Validation}}
    \item{\code{plotPCA}:}{List containing the parameters of the function \link{plotPCA}}
    \item{\code{plotPLS}:}{List containing the parameters of the function \link{plotPLS}}
    \item{\code{plotHeatmap}:}{List containing the parameters of the function \link{plotHeatmap}}
  }
}
\author{Francesc Fernandez,
 \email{francesc.fernandez.albert@upc.edu}}
\seealso{
\code{\link{MAIT-class}}}  
\keyword{classes}
