\name{project}
\alias{project}
\title{Change the basis of the MAIT data}
\description{
Function \code{project} is used to project the data of a MAIT object to the subspace of the models generated by another MAIT object.
}
\usage{

project(modelData,
        projectData)
}
\arguments{
  \item{modelData}{
The \code{\link{MAIT-class}} object where the models to which the new data is to be projected are saved.
}
\item{projectData}{
The \code{\link{MAIT-class}} containing the data to be projected.
}

}
\value{
	A matrix containing the data contained in the \code{projectData} parameter already projected into the \code{modelData} model subspace.
}
\author{Francesc Fernandez,
 \email{francesc.fernandez.albert@upc.edu}}
\keyword{file}
\keyword{internal}