\name{writeExcelTable}
\alias{writeExcelTable}
\title{Writes a csv table}
\description{
Function \code{writeExcelTable} writes a csv table with the input data.
}
\usage{

writeExcelTable(file,
                file.name)

}

\arguments{
  \item{file}{
The data to be saved in the csv file
}

  \item{file.name}{
The name of the csv file.
}
}

\value{
A csv file containg the data provided as input in the file parameter.
}

\author{Francesc Fernandez,
 \email{francesc.fernandez.albert@upc.edu}}
\keyword{internal}